WIP: Issue-399
The issue is that ng-change and input directive can be registered in any order.
This makes it possible for on-change event to execute before input event gets a
chance to run.

Changing from ng-change to on-change fixes the issue, since events are on-change
has to bubble to EventHandler before it can run. This is not quite right, since if
EventHandler is registered on the same element as input element, then again we don't
know which event listener will run first.

The separate registration model also means that apply() gets run for each element,
which is wasteful.

A proper way to fix this is for EventHandler to handle both the change event for
both on-change as well as for input element. This means that EventHandler needs to
be able to call listener function on InputElement. The only way to do that is for the
EventHandler to attach the listener function to the Expando. Currently we are favoring
the NgElement over NgPorbe object. So NgElement needs to have a list of listeners from
underlying directives which are not expressions on the DOM (as is on-change). These
listeners need to be stored in NgElement and retrieved by the EventHandler when event
fires.