chore: upgrade to barback 0.13.x
Pub in 1.4 is requiring barback 0.13 which has a breaking change in isPrimary (see https://codereview.chromium.org//223553008). This change allows Angular to target either barback 0.11.1+ or 0.13.0, depending on pub version.

End result is that this allows Angular to work on both stable and dev channel of Dart.

The corresponding DI change is https://github.com/angular/di.dart/pull/83

Closes #892